# src/main/resources/static/openapi.yaml
openapi: 3.1.0

info:
  title: CleaningApp API
  description: API for users, households, tasks, privileges, transactions and activity management.
  version: 1.0.0

servers:
  - url: 'http://localhost:8080/api'
    description: Local development server

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        firebaseUid:
          type: string
        name:
          type: string
        email:
          type: string
        createdAt:
          type: string
          format: date-time
        avatarUrl:
          type: string
          nullable: true

    UserRegisterRequest:
      type: object
      required:
        - name
        - email
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 30
        email:
          type: string
          format: email
        avatarUrl:
          type: string
          nullable: true

    ExcResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
          nullable: true
        time:
          type: string
          format: date-time

    HouseholdResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        inviteCode:
          type: string
        createdAt:
          type: string
          format: date-time
        createdByUser:
          type: string
          format: uuid
        isActive:
          type: boolean

    HouseholdRegisterRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string

    UserHouseholdJoinRequest:
      type: object
      properties:
        inviteCode:
          type: string
      required:
        - inviteCode

    UserHouseholdResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        householdId:
          type: string
          format: uuid
        balance:
          type: integer
        joinedAt:
          type: string
          format: date-time
        isUserActive:
          type: boolean

    TaskResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        household_id:
          type: string
          format: uuid
        created_by:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        title:
          type: string
        description:
          type: string
          nullable: true
        reward:
          type: integer
        assigned_to:
          type: string
          format: uuid
          nullable: true
        assigned_at:
          type: string
          format: date-time
          nullable: true
        is_completed:
          type: boolean
        completed_by:
          type: string
          format: uuid
          nullable: true
        completed_at:
          type: string
          format: date-time
          nullable: true

    TaskCreateRequest:
      type: object
      required:
        - title
        - reward
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        reward:
          type: integer
          minimum: 5
          maximum: 100

    PrivilegeResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        household_id:
          type: string
          format: uuid
        created_by:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        title:
          type: string
        cost:
          type: integer
        is_active:
          type: boolean

    PrivilegeCreateRequest:
      type: object
      required:
        - title
        - cost
      properties:
        title:
          type: string
        cost:
          type: integer

    TransactionResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        household_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        amount:
          type: integer
        type:
          type: string
          enum: [TASK_COMPLETION, PRIVILEGE_BOUGHT]
        related_entity_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time

    ActivityResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        household_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
          nullable: true
        activity_type:
          type: string
          enum: [TASK_CREATED, TASK_ASSIGNED, TASK_COMPLETED, PRIVILEGE_CREATED, PRIVILEGE_BOUGHT, USER_JOINED, USER_LEFT, HOUSEHOLD_CREATED]
        title:
          type: string
        related_entity_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time

security:
  - BearerAuth: []

paths:
  # -----------------------------
  # Auth
  # -----------------------------
  /auth/register:
    post:
      summary: Регистрация пользователя
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegisterRequest'
      responses:
        '201':
          description: Успешная регистрация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '409':
          description: USER_ALREADY_EXISTS или EMAIL_ALREADY_USED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExcResponse'

  # -----------------------------
  # User profile
  # -----------------------------
  /users/me:
    get:
      summary: Получение профиля пользователя
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Информация о пользователе
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: USER_NOT_FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExcResponse'
    put:
      summary: Редактирование профиля пользователя
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegisterRequest'
      responses:
        '200':
          description: Обновленный профиль пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: USER_NOT_FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExcResponse'
        '409':
          description: EMAIL_ALREADY_USED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExcResponse'

    delete:
      summary: Удаление (деактивация) аккаунта текущего пользователя
      description: >
        Деактивирует аккаунт пользователя.
        Пользователь теряет доступ к системе, данные сохраняются.
      security:
        - BearerAuth: []
      responses:
        '204':
          description: Успешно
        '403':
          description: USER_NOT_ACTIVE или ACCESS_DENIED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExcResponse'
        '404':
          description: USER_NOT_FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExcResponse'

  # -----------------------------
  # Households
  # -----------------------------
  /households:
    post:
      summary: Создать домохозяйство
      security:
        - BearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HouseholdRegisterRequest'

      responses:
        '201':
          description: Домохозяйство создано
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HouseholdResponse'
        '409':
          description: BUSINESS_CONFLICT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExcResponse'
        '403':
          description: ACCESS_DENIED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExcResponse'

    get:
      summary: Получить список домохозяйств пользователя
      security:
        - BearerAuth: [ ]
      responses:
        '200':
          description: Список домохозяйств
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HouseholdResponse'

  /households/join:
    post:
      summary: Присоединение к домохозяйству по коду
      security:
        - BearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserHouseholdJoinRequest'
      responses:
        '200':
          description: Пользователь присоединился к хозяйству
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserHouseholdResponse'
        '404':
          description: HOUSEHOLD_NOT_FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExcResponse'
        '403':
          description: HOUSEHOLD_NOT_ACTIVE или MEMBERSHIP_NOT_ACTIVE
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExcResponse'
        '409':
          description: BUSINESS_CONFLICT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExcResponse'

  /households/{householdId}:
    get:
      summary: Получить информацию о домохозяйстве
      security:
        - BearerAuth: [ ]
      parameters:
        - name: householdId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Информация о домохозяйстве
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HouseholdResponse'
        '404':
          description: HOUSEHOLD_NOT_FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExcResponse'
        '403':
          description: HOUSEHOLD_NOT_ACTIVE или MEMBERSHIP_NOT_ACTIVE
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExcResponse'
        '409':
          description: BUSINESS_CONFLICT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExcResponse'
    delete:
      summary: Удалить домохозяйство
      security:
        - BearerAuth: [ ]
      parameters:
        - name: householdId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Домохозяйство успешно удалено
        '404':
          description: HOUSEHOLD_NOT_FOUND или MEMBERSHIP_NOT_FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExcResponse'
        '403':
          description: HOUSEHOLD_NOT_ACTIVE или MEMBERSHIP_NOT_ACTIVE
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExcResponse'
    put: #/ надо оно ваще или нет этот патч
      summary: Редактирование хозяйства (имя)
      security:
        - BearerAuth: [ ]
      parameters:
        - name: householdId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HouseholdRegisterRequest'
      responses:
        '200':
          description: Обновленное домохозяйство
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HouseholdResponse'
        '404':
          description: HOUSEHOLD_NOT_FOUND или MEMBERSHIP_NOT_FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExcResponse'
        '403':
          description: HOUSEHOLD_NOT_ACTIVE или MEMBERSHIP_NOT_ACTIVE
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExcResponse'

  /households/{householdId}/leave:
    delete:
      summary: Покинуть домохозяйство
      security:
        - BearerAuth: [ ]
      parameters:
        - name: householdId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Пользователь покинул хозяйство
        '404':
          description: HOUSEHOLD_NOT_FOUND или MEMBERSHIP_NOT_FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExcResponse'
        '403':
          description: HOUSEHOLD_NOT_ACTIVE или MEMBERSHIP_NOT_ACTIVE
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExcResponse'

  /api/households/{householdId}/members:
    get:
      summary: Получить участников хозяйства
      security:
        - BearerAuth: []
      parameters:
        - name: householdId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Список участников
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
        '404':
          description: HOUSEHOLD_NOT_FOUND или MEMBERSHIP_NOT_FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExcResponse'
        '403':
          description: HOUSEHOLD_NOT_ACTIVE или MEMBERSHIP_NOT_ACTIVE
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExcResponse'

  /households/{householdId}/members/{userId}:
    delete:
      summary: Удалить участника из хозяйства
      description: "Любой участник может удалить другого участника (например, если тот удалил приложение)."
      security:
        - BearerAuth: [ ]
      parameters:
        - name: householdId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Участник удален
        '404':
          description: HOUSEHOLD_NOT_FOUND или MEMBERSHIP_NOT_FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExcResponse'
        '403':
          description: HOUSEHOLD_NOT_ACTIVE или MEMBERSHIP_NOT_ACTIVE
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExcResponse'
        '409':
          description: BUSINESS_CONFLICT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExcResponse'
# -----------------------------
# Tasks, Privileges, Transactions, Activity, Leaderboard
# (аналогично прописать CRUD эндпоинты для всех сущностей)
# -----------------------------

# Для сокращения здесь, могу сразу сделать полную версию с Task, Privilege, Transaction, Activity, Leaderboard